cleanup javascript serializer

+++ Explorers/Models/Achievements.cs
@@ -8,10 +8,10 @@ namespace WowDotNetAPI.Explorers.Models
 {
     public class Achievements
     {
-        public List<long> achievementsCompleted { get; set; }
+        public List<int> achievementsCompleted { get; set; }
         public List<long> achievementsCompletedTimestamp { get; set; }
-        public List<long> criteria { get; set; }
-        public List<long> criteriaQuantity { get; set; }
+        public List<int> criteria { get; set; }
+        public List<int> criteriaQuantity { get; set; }
         public List<long> criteriaTimestamp { get; set; }
         public List<long> criteriaCreated { get; set; }
     }
